Refund transactions are now imported as negative transfers.
Refund transactions are now imported as a negative transfer
of money from the asset account "to" the expense account.
Originally, they would have been imported as a transfer from
the expense account to the asset account.

This allows for refunds to be correctly accounted for
in progress and balance impacts. See the github wiki page
for more information on this decision.