Added unique ID to estimate class. Closes #101.
After assignments had been made, the actuals for an
estimate were retrieved from a map with the estimate
as an index. If the definition of the estimate changed
(i.e., the user changed the estimated amount), the actuals
were no longer accessible as the index had changed. Now,
the estimate ID is used for indexing and accessing actuals.
This means the user can change anything about an estimate
and its actuals will still be retrieved.