Budgeting period displayed correctly in form. Closes #116.
Added a budgeting period type parameter to budgeting periods,
allowing the edit-budget dialog to display the current period
upon launching the dialog. Also upon switching the period type,
the newly selected type is set as the current period. Previously,
after selecting a type, something had to be changed within that
type to force an update on the budget definition.