Added expected amount calculations. Closes #103.
What has not been implemented is automatically considering
estimate as final when the budget has expired (is in the past).
In this scenario, all expected values are the actual values,
and the entire expected column is identical to the actual
column. Thus it is completely redundant and information is
actually lost (can't retrieve expected value calculations).