Added context menu to create top-level estimates. Closes #110.
Also fixed a subtle bug in which the associated rules of an estimate
are not retrieved correctly after an estimate is modified. This is
because the old implementation of equals() in the estimate would
be false (as it should), but the estimate itself was the lookup
key for the rules. The estimate ID is now used for rule lookup.