Fix issue when reordering estimates under same parent.
The index that was stored when moving an estimate to the end of the
list of a parent's child estimates was not the correct index when
undoing the command (the total number of children didn't change).
Debug statements have been left in (for debug builds) because there
is still an issue after doing/undoing/redoing several move operations.