Restrict condition editing options based on field.
The options available in the drop-down combobox for the condition
operator will be restricted to only those applicable for the
currently selected condition field. Additionally, the case-sensitivity
checkbox will only be enabled for string type fields.

All "magic numbers" in the AssignmentRulesModel for the columns have
been replaced with file-scope constants.

Resolve issue #154.