Improve look of toolbar
The previous implementation used multiple toolbars for file, edit,
and analysis actions. Since none of these toolbars were moveable
or floatable, there was little reason to keep them as separate
toolbars. By moving them to a single toolbar and adding separators
between the action groups, the visible separation is clearer
(especially on look-and-feels that did not display a verticle line
between toolbars).