Disable clang in CI build
Travis-CI now uses clang 3.4--which now checks for friend declarations
with default arguments (must be definition). This was previously
unchecked by clang and gcc, even though it is part of the C++ spec.
Qt 5.2 has resolved this violation, however only Qt 5.0 is easily
acquired in the CI environment (Ubuntu 12.04).

Clang should be re-enabled if the CI environment ever reaches
a compatible combination.