Add budget version restriction to XML schema
This will allow the schema to validate only version 5.0 and above.
As the budget format changes, this schema should be updated to
maintain compliance.

If a non-backwards compatible change is introduced (i.e., version
6.0), then the schema needs to be updated for that change and the
allowable version increased to 6.0. Anyone wishing to validate 5.x
would need to obtain an older copy of the schema.